+++ CODE-mp/renderer/tr_public.h
@@ -42,7 +42,7 @@ typedef struct {
 	// size display elements
 	void	(*BeginRegistration)( glconfig_t *config );
 	qhandle_t (*RegisterModel)( const char *name );
-	int		  (*GetMd3Verts)( qhandle_t handle, const char* surfaceName, meshTriangle_t* meshTris, size_t bufferSizeTris );
+	int		  (*GetMd3Verts)( qhandle_t handle, const char* surfaceName, meshTriangle_t* meshTris, int bufferSizeTris );
 	qhandle_t (*RegisterSkin)( const char *name );
 	qhandle_t (*RegisterShader)( const char *name );
 	qhandle_t (*RegisterShaderNoMip)( const char *name );
